requirements.txt
-i https://pypi.org/simple
asn1crypto==0.24.0
bs4
cached-property==1.5.1
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
click==7.0
cryptography==2.3.1
defusedxml==0.5.0
dnspython==1.15.0
exchangelib
flask==1.0.2
future==0.16.0
idna==2.7
isodate==0.6.0
itsdangerous==0.24
jinja2==2.10
lxml==4.2.5
markupsafe==1.0
ntlm-auth
pycparser==2.19
pygments==2.2.0
python-dateutil
python-magic==0.4.15
pytz
requests-ntlm
requests
six==1.11.0
python-slugify==1.2.6
thehive4py
tzlocal
urllib3
werkzeug==0.14.1
hurry.filesize
